Committing Final Changes
Unless I work on this if I get to school early, this will be my last commit for the final project.

Tidied up a bit more, also added another scene to end the demo.

Also fixed a bug where the sprites in the treasure chests were not showing over the treasure chest.

Added a turret enemy to the overworld after defeating the dungeon.